[llvm][NFC] Move LLVM passes into their own library
Purely because I think they fit better as their own module rather than part of the VM module
