+++ src/jllvm/CMakeLists.txt
@@ -14,6 +14,7 @@
 add_subdirectory(class)
 add_subdirectory(compiler)
 add_subdirectory(gc)
+add_subdirectory(llvm)
 add_subdirectory(main)
 add_subdirectory(materialization)
 add_subdirectory(object)
